Optimized bins implementation of sen_slope
Use the estimates from the sample to also predict which bins will
contain median slope, lcl, ucl. This avoids one re-counting,
if the original estimates are close enough to the actual bin sizes
to not cause a bin to overflow or to have chosen the wrong bin(s).

Also minor change in error reporting.